Revert "Created a profile screen that allows users to edit their name, and job description. The profile image is yet to be completed, but you can change your image but cannot display it yet. on the profile page, you can see info about you as well as on the more page."
